Add csv files with all products and stock.
