+++ app/code/Webjump/IBCBackend/Console/Command/AddProductsCommand.php
@@ -4,7 +4,7 @@ namespace Webjump\IBCBackend\Console\Command;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
-use Webjump\IBCBackend\Model\Product\AddProducts;
+use Webjump\IBCBackend\Model\Product\Importer;
 use Magento\Framework\Console\Cli;
 
 Class AddProductsCommand extends Command
@@ -12,9 +12,9 @@ Class AddProductsCommand extends Command
     const INPUT_KEY_NAME = 'name';
     const INPUT_KEY_DESCRIPTION = 'description';
 
-    private AddProducts $itemFactory;
+    private Importer $itemFactory;
 
-    public function __construct(AddProducts $itemFactory)
+    public function __construct(Importer $itemFactory)
     {
         $this->itemFactory = $itemFactory;
         parent::__construct();
@@ -28,8 +28,7 @@ Class AddProductsCommand extends Command
 
     protected function execute(InputInterface $input, OutputInterface $output)
     {
-        $item = $this->itemFactory->importProducts();
-        var_dump($item);
+        $item = $this->itemFactory->execute();
         return Cli::RETURN_SUCCESS;
     }
 }
