+++ app/code/Webjump/IBCBackend/Model/Product/csv/.~lock.productsGames.csv#
@@ -1,0 +1,1 @@
+,webjump252,webjump252-Inspiron-3501,03.09.2021 15:54,file:///home/webjump252/.config/libreoffice/4;
