+++ app/code/Webjump/IBCBackend/Model/Product/csv/productsGames.csv
@@ -1,0 +1,14 @@
+sku,store_view_code,attribute_set_code,product_type,categories,product_websites,name,description,short_description,weight,product_online,tax_class_name,visibility,price,special_price,special_price_from_date,special_price_to_date,url_key,meta_title,meta_keywords,meta_description,base_image,base_image_label,small_image,small_image_label,thumbnail_image,thumbnail_image_label,swatch_image,swatch_image_label,created_at,updated_at,new_from_date,new_to_date,display_product_options_in,map_price,msrp_price,map_enabled,gift_message_available,custom_design,custom_design_from,custom_design_to,custom_layout_update,page_layout,product_options_container,msrp_display_actual_price_type,country_of_manufacture,additional_attributes,qty,out_of_stock_qty,use_config_min_qty,is_qty_decimal,allow_backorders,use_config_backorders,min_cart_qty,use_config_min_sale_qty,max_cart_qty,use_config_max_sale_qty,is_in_stock,notify_on_stock_below,use_config_notify_stock_qty,manage_stock,use_config_manage_stock,use_config_qty_increments,qty_increments,use_config_enable_qty_inc,enable_qty_increments,is_decimal_divided,website_id,related_skus,related_position,crosssell_skus,crosssell_position,upsell_skus,upsell_position,additional_images,additional_image_labels,hide_from_product_page,custom_options,bundle_price_type,bundle_sku_type,bundle_price_view,bundle_weight_type,bundle_values,bundle_shipment_type,associated_skus,downloadable_links,downloadable_samples,configurable_variations,configurable_variation_labels
+JG-AV-GTA-1,,Games,simple,"Default Category,Default Category/Aventura",games_ibc,Grand Theft Auto V  - Xbox One,,,0.330000,1,0,"Catalog, Search",153.900000,,,,grand-theft-auto-v-xbox-one,Grand Theft Auto V  - Xbox One,Grand Theft Auto V  - Xbox One,Grand Theft Auto V  - Xbox One ,,,,,,,,,03/09/2021 15:11,03/09/2021 15:17,,,Block after Info Column,,,,Use config,,,,,Product -- Full Width,,Use config,,"age_rating=16,multiplayer=Yes",100.0000,0.0000,1,0,0,1,1.0000,1,10000.0000,1,1,1.0000,1,1,1,1,1.0000,1,0,0,0,,,,,,,,,,,,,,,,,,,,,
+JG-AV-RDR-1,,Games,simple,"Default Category,Default Category/Aventura",games_ibc,Red Dead Redemption 2 - Xbox One,,,0.330000,1,0,"Catalog, Search",167.800000,,,,grand-theft-auto-v-xbox-one-1,,,,,,,,,,,,03/09/2021 15:11,03/09/2021 15:18,,,Block after Info Column,,,,Use config,,,,,Product -- Full Width,,Use config,,"age_rating=16,multiplayer=Yes",100.0000,0.0000,1,0,0,1,1.0000,1,10000.0000,1,1,1.0000,1,1,1,1,1.0000,1,0,0,0,,,,,,,,,,,,,,,,,,,,,
+JG-SM-M-1,,Games,simple,"Default Category,Default Category/Simulação",games_ibc,Minecraft - Xbox One,,,0.330000,1,0,"Catalog, Search",63.070000,,,,grand-theft-auto-v-xbox-one-2,,,,,,,,,,,,03/09/2021 15:11,03/09/2021 15:11,,,Block after Info Column,,,,Use config,,,,,Product -- Full Width,,Use config,,"age_rating=3,multiplayer=Yes",100.0000,0.0000,1,0,0,1,1.0000,1,10000.0000,1,1,1.0000,1,1,1,1,1.0000,1,0,0,0,,,,,,,,,,,,,,,,,,,,,
+JG-LT-MK11-1,,Games,simple,"Default Category,Default Category/Luta",games_ibc,Mortal Kombat 11 - Xbox One,,,0.330000,1,0,"Catalog, Search",251.900000,,,,grand-theft-auto-v-xbox-one-3,,,,,,,,,,,,03/09/2021 15:11,03/09/2021 15:11,,,Block after Info Column,,,,Use config,,,,,Product -- Full Width,,Use config,,"age_rating=18,multiplayer=Yes",100.0000,0.0000,1,0,0,1,1.0000,1,10000.0000,1,1,1.0000,1,1,1,1,1.0000,1,0,0,0,,,,,,,,,,,,,,,,,,,,,
+JG-LT-UFC-1,,Games,simple,"Default Category,Default Category/Luta",games_ibc,UFC 3 - Xbox One,,,0.330000,1,0,"Catalog, Search",99.900000,,,,grand-theft-auto-v-xbox-one-4,,,,,,,,,,,,03/09/2021 15:11,03/09/2021 15:11,,,Block after Info Column,,,,Use config,,,,,Product -- Full Width,,Use config,,"age_rating=12,multiplayer=Yes",100.0000,0.0000,1,0,0,1,1.0000,1,10000.0000,1,1,1.0000,1,1,1,1,1.0000,1,0,0,0,,,,,,,,,,,,,,,,,,,,,
+JG-LT-IG-1,,Games,simple,"Default Category,Default Category/Luta",games_ibc,Injustice Gods Among Us - Xbox 360,,,0.330000,1,0,"Catalog, Search",249.900000,,,,grand-theft-auto-v-xbox-one-5,,,,,,,,,,,,03/09/2021 15:11,03/09/2021 15:11,,,Block after Info Column,,,,Use config,,,,,Product -- Full Width,,Use config,,"age_rating=16,multiplayer=Yes",100.0000,0.0000,1,0,0,1,1.0000,1,10000.0000,1,1,1.0000,1,1,1,1,1.0000,1,0,0,0,,,,,,,,,,,,,,,,,,,,,
+JG-ES-BR-1,,Games,simple,"Default Category,Default Category/Estratégia",games_ibc,Battlefield Revolution - Xbox One,,,0.330000,1,0,"Catalog, Search",88.990000,,,,grand-theft-auto-v-xbox-one-6,,,,,,,,,,,,03/09/2021 15:11,03/09/2021 15:11,,,Block after Info Column,,,,Use config,,,,,Product -- Full Width,,Use config,,"age_rating=16,multiplayer=Yes",100.0000,0.0000,1,0,0,1,1.0000,1,10000.0000,1,1,1.0000,1,1,1,1,1.0000,1,0,0,0,,,,,,,,,,,,,,,,,,,,,
+JG-TE-BW-1,,Games,simple,"Default Category,Default Category/Terror",games_ibc,Blair Witch - Xbox One,,,0.330000,1,0,"Catalog, Search",284.980000,,,,grand-theft-auto-v-xbox-one-7,,,,,,,,,,,,03/09/2021 15:11,03/09/2021 15:11,,,Block after Info Column,,,,Use config,,,,,Product -- Full Width,,Use config,,"age_rating=18,multiplayer=Yes",100.0000,0.0000,1,0,0,1,1.0000,1,10000.0000,1,1,1.0000,1,1,1,1,1.0000,1,0,0,0,,,,,,,,,,,,,,,,,,,,,
+JG-TE-RE-1,,Games,simple,"Default Category,Default Category/Terror",games_ibc,Resident Evil 3 - Xbox One,,,0.330000,1,0,"Catalog, Search",138.680000,,,,grand-theft-auto-v-xbox-one-8,,,,,,,,,,,,03/09/2021 15:11,03/09/2021 15:11,,,Block after Info Column,,,,Use config,,,,,Product -- Full Width,,Use config,,"age_rating=18,multiplayer=Yes",100.0000,0.0000,1,0,0,1,1.0000,1,10000.0000,1,1,1.0000,1,1,1,1,1.0000,1,0,0,0,,,,,,,,,,,,,,,,,,,,,
+JG-ES-TH-1,,Games,simple,"Default Category,Default Category/Esporte",games_ibc,Tony Hawk's Pro Skater 1 + 2 - Xbox One,,,0.330000,1,0,"Catalog, Search",329.900000,,,,tony-hawk-s-pro-skater-1-2-xbox-one,Tony Hawk's Pro Skater 1 + 2 - Xbox One,Tony Hawk's Pro Skater 1 + 2 - Xbox One,Tony Hawk's Pro Skater 1 + 2 - Xbox One ,,,,,,,,,03/09/2021 15:11,03/09/2021 15:11,,,Block after Info Column,,,,Use config,,,,,Product -- Full Width,,Use config,,"age_rating=7,multiplayer=No",100.0000,0.0000,1,0,0,1,1.0000,1,10000.0000,1,1,1.0000,1,1,1,1,1.0000,1,0,0,0,,,,,,,,,,,,,,,,,,,,,
+JG-ES-F-1,,Games,simple,"Default Category,Default Category/Esporte",games_ibc,FIFA 20 - Xbox One,,,0.330000,1,0,"Catalog, Search",129.900000,,,,tony-hawk-s-pro-skater-1-2-xbox-one-1,,,,,,,,,,,,03/09/2021 15:11,03/09/2021 15:11,,,Block after Info Column,,,,Use config,,,,,Product -- Full Width,,Use config,,"age_rating=7,multiplayer=Yes",100.0000,0.0000,1,0,0,1,1.0000,1,10000.0000,1,1,1.0000,1,1,1,1,1.0000,1,0,0,0,,,,,,,,,,,,,,,,,,,,,
+JG-SM-ETS-1,,Games,simple,"Default Category,Default Category/Simulação",games_ibc,Euro Truck Simulator 2 - Xbox One,,,0.330000,1,0,"Catalog, Search",219.000000,,,,the-sims-4-xbox-one,The Sims 4 - Xbox One,The Sims 4 - Xbox One,The Sims 4 - Xbox One ,,,,,,,,,03/09/2021 15:11,03/09/2021 15:11,,,Block after Info Column,,,,Use config,,,,,Product -- Full Width,,Use config,,"age_rating=7,multiplayer=No",100.0000,0.0000,1,0,0,1,1.0000,1,10000.0000,1,1,1.0000,1,1,1,1,1.0000,1,0,0,0,,,,,,,,,,,,,,,,,,,,,
+JG-SM-TS4,,Games,simple,"Default Category,Default Category/Simulação",games_ibc,The Sims 4 - Xbox One,,,0.330000,1,0,"Catalog, Search",106.070000,,,,the-sims-4-xbox-one-1,,,,,,,,,,,,03/09/2021 15:11,03/09/2021 15:16,,,Block after Info Column,,,,Use config,,,,,Product -- Full Width,,Use config,,"age_rating=12,multiplayer=No",100.0000,0.0000,1,0,0,1,1.0000,1,10000.0000,1,1,1.0000,1,1,1,1,1.0000,1,0,0,0,,,,,,,,,,,,,,,,,,,,,
