+++ app/code/Webjump/IBCBackend/Model/Product/csv/stockGames.csv
@@ -1,0 +1,32 @@
+source_code,sku,status,quantity
+default,JG-SM-GTA1,1,100.0000
+default,JG-SM-GTA1-1,1,100.0000
+default,JG-SM-RDR1,1,100.0000
+default,JG-SM-RDR1-1,1,100.0000
+default,JG-SM-M1,1,100.0000
+default,JG-SM-M1-1,1,100.0000
+default,JG-LT-MK11,1,100.0000
+default,JG-LT-MK11-1,1,100.0000
+default,JG-LT-UFC1,1,100.0000
+default,JG-LT-UFC1-1,1,100.0000
+default,JG-LT-IGAU,1,100.0000
+default,JG-LT-IGAU-1,1,100.0000
+default,JG-ES-BR,1,100.0000
+default,JG-ES-BR-1,1,100.0000
+default,JG-TE-BW,1,100.0000
+default,JG-TE-BW-1,1,100.0000
+default,JG-TE-RE-1,1,100.0000
+default,JG-TE-RE-1-1,1,100.0000
+default,JG-TE-BW-1-1,1,100.0000
+default,JG-SM-RDR-1,1,100.0000
+default,JG-SM-M-1,1,100.0000
+default,JG-SM-GTA-1,1,100.0000
+default,JG-LT-UFC-1,1,100.0000
+default,JG-LT-IG-1,1,100.0000
+default,JG-ES-TH-1,1,100.0000
+default,JG-ES-TH-1-1,1,100.0000
+default,JG-ES-F-1,1,100.0000
+default,JG-SM-ETS-1,1,100.0000
+default,JG-SM-TS4,1,100.0000
+default,JG-AV-GTA-1,1,100.0000
+default,JG-AV-RDR-1,1,100.0000
