+++ app/code/Webjump/IBCBackend/Model/Product/csv/stockSkate.csv
@@ -1,0 +1,29 @@
+default,SKT-SH-1,1,50.0000
+default,SKT-SH-1-1,1,50.0000
+default,SKT-SH-2,1,50.0000
+default,SKT-SH-2-1,1,50.0000
+default,SKT-SH-3,1,50.0000
+default,SKT-LX-1,1,40.0000
+default,SKT-LX-1-1,1,35.0000
+default,SKT-LX-1-1-1,1,35.0000
+default,SKT-LX-3,1,35.0000
+default,SKT-LX-2,1,35.0000
+default,ASDASDASD,1,123.0000
+default,"Roda Iron Long Hypnotic 83mm - Azul",1,50.0000
+default,"Roda Iron Long Hypnotic 83mm - Azul-1",1,50.0000
+default,SKT-RO-1,1,50.0000
+default,SKT-RO-1-1,1,50.0000
+default,SKT-RO-3,1,50.0000
+default,SKT-RO-2,1,50.0000
+default,SKT-CO-1,1,30.0000
+default,SKT-CO-1-1,1,30.0000
+default,SKT-CO-2,1,30.0000
+default,SKT-CO-2-1,1,30.0000
+default,SKT-CO-3,1,30.0000
+default,SKT-TK-1,1,40.0000
+default,SKT-TK-2,1,35.0000
+default,SKT-TK-2-1,1,35.0000
+default,SKT-TK-3,1,35.0000
+default,SKT-AC-1,1,30.0000
+default,SKT-AC-1-1,1,30.0000
+default,SKT-AC-2,1,60.0000
