+++ app/code/Webjump/IBCBackend/Model/Product/csv/website-productsSkate.csv
@@ -1,0 +1,2 @@
+sku,store_view_code,attribute_set_code,product_type,name
+SKT-SH-1,,Skate,simple,TESTE DEFAULT
